requirements.txt
easydict
dateutils
python-dotenv
tqdm
numpy
pandas
matplotlib
opencv-python
pytorch-lightning
scipy
networkx
scikit-learn
mlflow
google-cloud-storage
google-cloud-vision
google-cloud-documentai
# Install manually
# torch==1.10.1
# torchvision==0.11.2
# torch-scatter -f https://data.pyg.org/whl/torch-1.10.1+cu102.html
# torch-sparse -f https://data.pyg.org/whl/torch-1.10.1+cu102.html
# torch-geometric